🎉 Init keyring
